Refactor: reorganize test suite structure and relocate tests to appropriate unit directories
