+++ tests/unit/src/core/database/test_advisory_locks.py
