+++ tests/unit/src/core/errors/test_error_handlers.py
