+++ tests/unit/src/core/patterns/test_singleton.py
