+++ tests/unit/src/core/redis/cache/test_redis_cache_lifecycle.py
