+++ tests/unit/src/core/redis/test_redis_lifecycle.py
