+++ tests/unit/src/main/test_lifespan.py
