+++ tests/unit/src/user/services/test_user_service.py
